mindcareai_pr/requirements.txt
# Core Agent + LLM Framework
langchain>=0.2.0
langgraph>=0.0.56
langchain-openai>=0.1.7

# Machine Learning (Emotion Model)
scikit-learn>=1.5.0
pandas>=2.2.2
numpy>=1.26.4
joblib>=1.4.2

# NLP Preprocessing
nltk>=3.9.1

# Config & Utilities
python-dotenv>=1.0.1
requests>=2.32.3
email-validator>=2.2.0
wikipedia
textblob
uvicorn